- fixed: [Email module - client] fixed an issue where the "read" flag would not be properly written into MySQL based datastorages with strict mode enabled (closes CN-413)

+++ src/corelib/php/library/Conjoon/Modules/Groupware/Email/Item/Model/Flag.php
@@ -139,7 +139,7 @@ class Conjoon_Modules_Groupware_Email_Item_Model_Flag extends Conjoon_Db_Table {
             return 0;
         }
 
-        $data = array('is_read' => (bool)$isRead);
+        $data = array('is_read' => (int)((bool)$isRead));
         $adapter = $this->getAdapter();
         return $this->update($data, array(
             $adapter->quoteInto('groupware_email_items_id = ?', $groupwareEmailItemsId, 'INTEGER'),
